rhcos: Install via repos and rpm-ostree
This mirrors https://github.com/coreos/fcos-derivation-example/pull/1
but for RHCOS.

However for now, in order to avoid entitlements in the dependency
chain, demo with centos.  This required some surprising hacks; I
took the `.repo` files out of the centos container, but those .repo
files depend on e.g. `$releasever = 8` but it's of course different
for us.  There's also an `$infra` variable that I inlined; the value
is `container`.