Begin restructure to support multiple examples
I think this repository will be more compelling if we have *multiple*
examples.

For example, I think the "build a Go binary" thing demonstrates
non-RPM content, but it also confuses things a bit because
many users will *not* want to do that.  They will want to see a simpler
flow that e.g. just layers some packages and adds configuration.

Move the current one to a subdirectory in preparation for multiple.