generate-rhcos-versions: Only inject `ADD rhel.repo` if needed
Only some of the examples actually install packages, not all.
In particular the `selinux` one does not, and I want to use it
as a test case to fix https://github.com/openshift/os/issues/829

I'm thinking actually we should drop this bit entirely and rely
on host subscriptions *or* users doing it on their own, but
for now this works.